pdating api docs (swagger) removed sbtc.world refs updated for stacks-network repo change. updated for romeo contract in api/web
