+++ sbtc-bridge-api/README.md
@@ -1,26 +1,3 @@
 # sbtc-bridge-api
 
-![ci](https://github.com/Trust-Machines/sbtc-bridge-api)
-
-## Introduction
-
-Node Server for sBTC Bridge.
-
-[sbtc.world](https://sbtc.world).
-
-## Build
-
-```bash
-npm install
-npm run build
-```
-
-## Test
-
-```bash
-npm run test
-```
-
-## Deploy
-
-To be integrated with Google Cloud. Currently builds to Linode via ssh and rsync.
+see main readme file.
