+++ sbtc-bridge-api/src/app.ts
@@ -31,7 +31,7 @@ app.get('/api-docs', swaggerUi.setup(swaggerDocument));
 setConfigOnStart();
 app.use(
   bodyParser.urlencoded({
-    extended: true,
+    extended: true, 
   })
 );
 app.use(bodyParser.json());
