+++ sbtc-bridge-api/src/controllers/ConfigController.ts
@@ -2,9 +2,12 @@ import { Get, Route } from "tsoa";
 import { getConfig } from '../lib/config.js';
 
 export interface IStringToStringDictionary { [key: string]: string|number|undefined; }
+
+
+
 @Route("/bridge-api/:network/v1/config")
 export class ConfigController {
-  @Get("/")
+  //@Get("/")
   public getAllParam(): any {
     const config = getConfig();
     config.mongoDbUrl = '*****'
@@ -16,7 +19,7 @@ export class ConfigController {
     return config;
   }
 
-  @Get("/:param")
+  //@Get("/:param")
   public getParam(param:string): string|number|undefined {
     let myval;
     for (const [key, value] of Object.entries(getConfig())) {
