+++ sbtc-bridge-api/src/routes/stacksRoutes.ts
@@ -5,6 +5,52 @@ import type { PeginRequestI } from 'sbtc-bridge-lib';
 
 const router = express.Router();
 
+router.get("/build/deposit/:stxAddress/:revealFee", async (req, res, next) => {
+  try {
+    const controller = new DepositsController();
+    const response = await controller.commitDepositData(req.params.stxAddress, Number(req.params.revealFee));
+    return res.send(response);
+  } catch (error) { 
+    console.log('Error in routes: ', error)
+    next('An error occurred fetching sbtc data.') 
+  }
+});
+
+router.get("/parse/deposit/:data", async (req, res, next) => {
+  try {
+    const controller = new DepositsController();
+    const response = await controller.commitDeposit(req.params.data);
+    return res.send(response);
+  } catch (error) {
+    console.log('Error in routes: ', error)
+    next('An error occurred fetching sbtc data.') 
+  }
+});
+
+router.get("/build/withdrawal/:signature/:amount", async (req, res, next) => {
+  try {
+    const controller = new DepositsController();
+    const response = await controller.commitWithdrawalData(req.params.signature, Number(req.params.amount));
+    return res.send(response);
+  } catch (error) {
+    console.log('Error in routes: ', error)
+    next('An error occurred fetching sbtc data.')
+  }
+});
+
+
+router.get("/parse/withdrawal/:data/:sbtcWallet", async (req, res, next) => {
+  try {
+    const controller = new DepositsController();
+    const response = await controller.commitWithdrawal(req.params.data, req.params.sbtcWallet);
+    return res.send(response);
+  } catch (error) {
+    console.log('Error in routes: ', error)
+    next('An error occurred fetching sbtc data.') 
+  }
+});
+
+
 router.get("/address/:address/balance", async (req, res, next) => {
   try {
     const controller = new SbtcWalletController();
