+++ sbtc-signer-web/src/lib/config.ts
@@ -29,7 +29,7 @@ const TESTNET_CONFIG = {
     VITE_PUBLIC_APP_VERSION: '1.0.0',
     VITE_URI_BRIDGE: 'https://bridge.stx.eco',
     VITE_URI_SIGN: 'https://bridge.stx.eco',
-    VITE_ORIGIN: 'https://sbtc.world',
+    VITE_ORIGIN: 'https://sbtc.tech',
     VITE_NETWORK: 'testnet',
     VITE_SBTC_WALLET: 'tb1q6zlpyrzvzjcrf6dlsctcrh9yl3dwfktsw0nclq',
     VITE_SBTC_CONTRACT_ID: 'ST306HDPY54T81RZ7A9NGA2F03B8NRGW6Y59ZRZSD.faint-tan-cobra',
@@ -51,7 +51,7 @@ const MAINNET_CONFIG = {
     VITE_PUBLIC_APP_VERSION: '1.0.0',
     VITE_URI_BRIDGE: 'https://bridge.stx.eco',
     VITE_URI_SIGN: 'https://bridge.stx.eco',
-    VITE_ORIGIN: 'https://sbtc.world',
+    VITE_ORIGIN: 'https://sbtc.tech',
     VITE_NETWORK: 'mainnet',
     VITE_SBTC_WALLET: 'tb1q6ue638m4t5knwxl4kwhwyuffttlp0ffee3zn3e',
     VITE_SBTC_CONTRACT_ID: 'ST3N4AJFZZYC4BK99H53XP8KDGXFGQ2PRSPNET8TN.sky-blue-elephant',
@@ -73,7 +73,7 @@ const DEVNET_CONFIG = {
     VITE_PUBLIC_APP_VERSION: '1.0.0',
     VITE_URI_BRIDGE: 'http://localhost:8080',
     VITE_URI_SIGN: 'http://localhost:8081',
-    VITE_ORIGIN: 'https://sbtc.world',
+    VITE_ORIGIN: 'https://sbtc.tech',
     VITE_NETWORK: 'devnet',
     VITE_SBTC_WALLET: 'tb1q6zlpyrzvzjcrf6dlsctcrh9yl3dwfktsw0nclq',
     VITE_SBTC_COORDINATOR: 'ST2BJA4JYFJ7SDMNFJZ9TJ3GB80P9Z80ADNF2R2AG',
